ref(VirtioDeviceActions): include default implementations
This patch is based on restructuring the implementations of
some functions of the VirtioDeviceActions trait as it now
includes default methods. In this case, now only the functions
strictly necessary for the device type must be implemented
(e.g. for devices of the vhost-user type, the read_config
function must be implemented if the device has negotiated
the CONFIG protocol feature).

Signed-off-by: joaopeixoto13 <joaopeixotooficial@gmail.com>