ref(event_manager): remote endpoint creation offload
This commit is based on the offload of the remote endpoint
object (the object that is capable of interacting with the
event manager), from the structure common to all VirtIO devices
(VirtioDeviceCommon struct) for each of the VirtIO devices.
This is because only devices that are implemented within the
VMM / Device Model (such as common VirtIO) need this object,
as devices such as vhost/vhost-user implement the handling
of guest-to-host notifications themselves.

Signed-off-by: joaopeixoto13 <joaopeixotooficial@gmail.com>