ref(ioctls): remove create and destroy I/O clients calls
This commit removes the creation and deletion ioctls
regarding I/O clients, since the VMM / Frontend DM
only interacts with the Control I/O Client. In this way,
the creation of this client is performed when the
backend device model is created by the
BAO_IOCTL_VM_VIRTIO_BACKEND_CREATE ioctl issued by the VMM.
To destroy the Control I/O Client, the VMM simply calls
the BAO_IOCTL_VM_VIRTIO_BACKEND_DESTROY ioctl which,
internally, it is called the delete method for all I/O
Clients (Control and Ioeventfd).

Signed-off-by: joaopeixoto13 <joaopeixotooficial@gmail.com>