ref(ioctl): removed create/delete ioctls and added get dm info
This commit removes the need of creating and deleting the device
model abstraction inside the kernel since now it is the reponsability
of the I/O Dispatcher system to manage that.
To retrive the device model information (e.g., shared memory region
and device model file descriptor), we created a dedicated ioctl.

Signed-off-by: João Peixoto <joaopeixotooficial@gmail.com>