fix(pty_handler): verify if pty is open in pty source event
This commit adds a check to ensure the pty slave is open on the
backend by any process (e.g., minicom, picocom) during the pty
source event. This prevents transmitting guest output messages
back to the frontend when the backend console is not open.

Note: This is a temporary workaround to ensure functionality
while investigating why the `SOURCE_PTY` event is not triggered
when the backend console is open.

Signed-off-by: João Peixoto <joaopeixotooficial@gmail.com>